add modal for adding in the data for leaderboard
